clean main

+++ Assets/Scripts/Score.cs
@@ -1,4 +1,3 @@
-using System;
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
@@ -13,6 +12,6 @@ public class Score : MonoBehaviour
     private void FixedUpdate()
     {
         _score += 1;
-        _scoreText.text = _score.ToString();
+        _scoreText.text = ((float)(_score)).ToString();
     }
 }
